requirements.txt
imutils==0.5.3
numpy==1.18.1
opencv-contrib-python==4.2.0.32
Pillow==7.0.0
pkg-resources==0.0.0
six==1.14.0
torch==1.4.0+cpu
torchvision==0.5.0+cpu
